refactor get_all_formes into utilities
